requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
EbookLib==0.17.1
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.4.2
newspaper3k==0.2.8
nltk==3.4.5
Pillow==6.2.1
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
requests-file==1.4.3
six==1.13.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
